Update and rename 003_list_less_than_ten.py to challenge-003-LessThanTen
Third challenge (that I gave myself), first to be put in a repository. 
Essentially this is a program that takes user input to populate a list of integers. After the user is finished creating the list, the program asks for another integer, which it then uses as a threshold, removing all integers larger than it from the list. After that has happened it sorts the remaining elements in ascending order and returns the list.