Update and rename stockInfo.py to stockrequest.py
I realize there are a *lot* of redundancies and seemingly pointless code here. I uploaded so as to have remote access tomorrow, for now I am going to sleep and stopping in the middle of trying to figure something new out. This is what new concepts look like in my code: very rough until I understand just what the hell is going on.